Override api_extensions_path in unit tests
In order to be able to implement unit tests for FWaaS, the extension
path must be set to the midonet's in setup.  The reason is that FWaaS
puts its extension path twice, one in the CONF override and one
inserted by Neutron separately because it's a well-known advanced
service.  This is usually not a problem because one of the duplicated
paths gets removed when Neutron starts, but this dedup logic fails as
soon as you append a different extension path, causing Neutron to fail
to start.  This patch sets the extension path of the midonet plugin in
the test setup so that any duplicated paths that might exist in CONF
get overwritten completely.

Partially-implements:
https://blueprints.launchpad.net/networking-midonet/+spec/midonet-fwaas
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>

Change-Id: I8e5d7c8d7ee793f63346fe856e782c22d829dfe4