Listen alternative topic for state reports
Recently Neutron agents have been changed to use topics.REPORTS
for state reports. [1]  Update our plugin to accept the topic as well.

In case of MidoNet, this topic is used to accept messages from
Neutron dhcp and metadata agents.  While MidoNet v2015.09 and
later will have an alternative metadata proxy [2] which doesn't use
these agents, it's nice to maintain the old way at least for
a cycle.  And v1 plugin needs this anyway.

This patch is required to run with MOS 7.0.

[1] Id86a1f962aaa4f64011d57ae55d240f890cca4f7
[2] https://github.com/midonet/midonet/blob/c44eadbf1bd676e232099487b81cfcf6189d3bba/docs/openstack-metadata-usage.md

Related-Bug: #1496410
Change-Id: I65aeb3497cc11e0603a8c7af9bae129088924935