tox_install: Fix "already installed" test
For many of projects, the name of projects are slightly different from
the name of modules.  eg. neutron_fwaas vs neutron-fwaas.

NOTE: This patch also updates test-requirements from cherry-picked
patch. Without this update, fixture 2.0 is used and it causes error,
though the reason why fixture 2.0 is used is unclear.

Conflicts:
	tools/tox_install.sh
	tools/tox_install_project.sh

Closes-Bug: #1514280
Change-Id: I9027ce89f09e85085ce5ce1c7d3da872c840b94f