logging_resource: Separate models from db mixins
So that db migration can be run without neutron-fwaas installed.

Note: v1 plugin doesn't support fwaas.

Note: the FK to firewalls.id should be ok because the table is
created by db migration of neutron. (not -fwaas)

Backport note: This change is not necessary for this branch
as for kilo midonet-db-manage doesn't import models in the first place.
But it's better to keep in sync with upstream master.

Closes-Bug: #1597604
Change-Id: I9a58f38449543700042e353a384fde75951f9a2f
(cherry picked from commit bb7f5d0b96038333407800e06efb92622007b84a)