FIP6: Support creating IPv6 FIP
This modifies l3_db_midonet to override the _create_floatingip
method in order to allow the creation of floating IPs with IPv6
addresses. The patch also fixes a bug in the overriden method
where if creating the floating IP will fail after creating the
port, the port was not deleted.

Ref: MI-1603

Change-Id: Ia852002aaad32e5c91c45ecf5f93a8c0c432c1b7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>